#0: Allow setting worker L1 size at device creation time
Currently we use a dispatcher ringbuffer that's 69 kB, which always works, but
has limited dispatch parallelism when programs have large kernels.

To help improve this, add a flag to the device so users can specify how much
worker L1 space they need to store the data. They can reduce this to increase
the size of the ringbuffer. If no value is specified, default is used which
will set the ringbuffer size to 69 kB.

With this change, the start of unreserved memory isn't a constant that's
exposed through hal, but should be retrieved from the allocator.

The ethernet core ringbuffer size stays fixed, since it's often it's used by
fabric, which is outside the application's normal control.